Loan

Documentation for Loan.

Loan.amount_paid

Loan.days_between

Loan.due_dates

Loan.factor_price

Loan.installment

Loan.installment

Loan.merge_loan

Loan.present_value

Loan.present_value

Loan.total_month_installment

Loan.value_at_date

Loan.value_at_date

Loan.value_by_pmt

Loan.amount_paid — Method

amount_paid(loan::LoanAgreement,installment::Installment,date::Dates.Date,payment::Pair{Dates.Date,Float64})

Return the amount already paid of a installment until a date.

Loan.days_between — Method

days_between(dates::AbstractVector{Dates.Date},initialDate::Dates.Date)

Loan.due_dates — Method

due_dates(initialDate::Dates.Date,nper::Number,calendar::Symbol;period=Dates.Month,grace=0)

Return nper of due dates after initialDate + grace period (in days) that is a business day in calendar

Loan.factor_price — Method

factor_price(rate,qtPeriods,qtPeriodsFirst)

.

Loan.installment — Method

installment(amount::Number,rate::Number,initialDate::Dates.Date,dueDates::AbstractVector{Dates.Date};period=Dates.Month)

Return the constant payment value required to settle a loan (amount) with a fixed rate with dueDates payments flow

Loan.installment — Method

installment(amount::Number,rate::Number,initialDate::Dates.Date,nper::Number,calendar::Union{Symbol,T}=BusinessDays.NullHolidayCalendar();period=Dates.Month) where T<: BusinessDays.HolidayCalendar

Return the constant payment value required to settle a loan (amount) with a fixed rate agreed at initialDate in nper payments

Loan.merge_loan — Method

merge_loan(a::LoanAgreement,b::LoanAgreement)

Merge two LoanAgreements a an a, resulting a new LoanAgreement which cash flow is equals to the sum of a and b.

Loan.present_value — Method

present_value(amount,rate,dueDate,presentDate;period=Dates.Month)

Return the value at presentDate of value of the amount due to dueDate discounted by the rate.

Loan.present_value — Method

present_value(amount,rate,dueDate;period=Dates.Month)

Return the present value of the amount due to dueDate discounted by the rate.

Loan.total_month_installment — Method

total_month_installment(loan::T, month::Dates.Date) where T<:LoanAgreement

The total value due of the loan by the month

Loan.value_at_date — Method

value_at_date(installment::Installment,date::Dates.Date)

Return the present value of the non-paid installment or the paid value before the date.

Loan.value_at_date — Method

value_at_date(loan::LoanAgreement,date::Dates.Date)

Return the present value of the non-paid plus the amount paid until the date of a loan agreement.

Loan.value_by_pmt — Method

value_by_pmt(pmt::Number,nper::Number,rate::Number,initialDate::Dates.Date,firstDueDate::Dates.Date;calendar=BusinessDays.NullHolidayCalendar(),period=Dates.Month)

Which value results in a regular payments of pmt if paid in n constant installments discounted by the rate.